+++ packages/rook_sdk_core/lib/rook_sdk_core.dart
@@ -1,14 +1,24 @@
 library rook_sdk_core;
 
+export 'src/enum/request_permissions_status.dart';
 export 'src/enum/rook_environment.dart';
 export 'src/enum/sync_status.dart';
+export 'src/exception/bad_user_timezone_exception.dart';
+export 'src/exception/connect_timeout_exception.dart';
+export 'src/exception/date_not_valid_for_events_exception.dart';
+export 'src/exception/date_not_valid_for_summaries_exception.dart';
+export 'src/exception/device_not_supported_exception.dart';
+export 'src/exception/health_connect_not_installed_exception.dart';
 export 'src/exception/http_request_exception.dart';
+export 'src/exception/missing_android_permissions_exception.dart';
 export 'src/exception/missing_configuration_exception.dart';
-export 'src/exception/missing_permissions_exception.dart';
+export 'src/exception/missing_health_connect_permissions_exception.dart';
+export 'src/exception/request_quota_exceeded_exception.dart';
 export 'src/exception/sdk_not_authorized_exception.dart';
 export 'src/exception/sdk_not_initialized_exception.dart';
-export 'src/exception/timeout_exception.dart';
+export 'src/exception/user_not_deleted_exception.dart';
 export 'src/exception/user_not_initialized_exception.dart';
+export 'src/exception/user_not_registered_exception.dart';
 export 'src/model/data_source.dart';
 export 'src/model/rook_configuration.dart';
 export 'src/model/sync_status_with_data.dart';