+++ packages/rook_sdk_core/lib/src/exception/bad_user_timezone_exception.dart
@@ -1,0 +1,11 @@
+/// Exception thrown when the provided timezone offset is out of range.
+class BadUserTimeZoneException implements Exception {
+  final String message;
+
+  BadUserTimeZoneException(this.message);
+
+  @override
+  String toString() {
+    return 'BadUserTimeZoneException{message: $message}';
+  }
+}