+++ packages/rook_sdk_core/lib/src/exception/date_not_valid_for_summaries_exception.dart
@@ -1,0 +1,11 @@
+/// Exception thrown when the provided date is not valid for summaries extraction.
+class DateNotValidForSummariesException implements Exception {
+  final String message;
+
+  DateNotValidForSummariesException(this.message);
+
+  @override
+  String toString() {
+    return 'DateNotValidForEventsException{message: $message}';
+  }
+}