+++ packages/rook_sdk_core/lib/src/exception/device_not_supported_exception.dart
@@ -1,3 +1,4 @@
+/// Exception thrown when the current device is not supported.
 class DeviceNotSupportedException implements Exception {
   final String message;