+++ packages/rook_sdk_core/lib/src/exception/http_request_exception.dart
@@ -1,12 +1,15 @@
 /// Exception thrown when a http request fails.
+///
+/// [code] The http status code.
+/// [error] The error message.
 class HttpRequestException implements Exception {
-  final String message;
   final int code;
+  final String error;
 
-  HttpRequestException(this.message, this.code);
+  HttpRequestException(this.code, this.error);
 
   @override
   String toString() {
-    return 'HttpRequestException{message: $message, code: $code}';
+    return 'HttpRequestException{code: $code, error: $error}';
   }
 }