+++ packages/rook_sdk_core/lib/src/exception/sdk_not_authorized_exception.dart
@@ -1,3 +1,4 @@
+/// Exception thrown when the SDK is not authorized.
 class SDKNotAuthorizedException implements Exception {
   final String message;