+++ packages/rook_sdk_health_connect/test/src/data/extension/result_sync_status_extensions_test.dart
@@ -3,9 +3,6 @@ import 'package:rook_sdk_core/rook_sdk_core.dart';
 import 'package:rook_sdk_health_connect/src/data/extension/result_boolean_extensions.dart';
 import 'package:rook_sdk_health_connect/src/data/extension/result_sync_status_extensions.dart';
 import 'package:rook_sdk_health_connect/src/data/proto/protos.pb.dart';
-import 'package:rook_sdk_health_connect/src/domain/exception/device_not_supported_exception.dart';
-import 'package:rook_sdk_health_connect/src/domain/exception/health_connect_not_installed_exception.dart';
-import 'package:rook_sdk_health_connect/src/domain/exception/request_quota_exceeded_exception.dart';
 
 void main() {
   group('ResultSyncStatusProto success', () {
@@ -73,8 +70,7 @@ void main() {
           throwsA(
             predicate(
               (exception) =>
-                  exception is HttpRequestException &&
-                  exception.message == error,
+                  exception is HttpRequestException && exception.error == error,
             ),
           ),
         );
@@ -93,7 +89,7 @@ void main() {
           throwsA(
             predicate(
               (exception) =>
-                  exception is MissingPermissionsException &&
+                  exception is MissingAndroidPermissionsException &&
                   exception.message == error,
             ),
           ),